requirements.txt
requests 
pandas 
tweepy 
newsapi-python 
alpha_vantage
yfinance
textblob
transformers 
torch 
nltk 
spacy 
scikit-learn
xgboost
keras
tensorflow
google-cloud-storage
google-cloud-bigquery
apache-airflow[gcp]